requirements.txt
fastapi[all]
supabase
strawberry-graphql[fastapi]
